FastCap2 Windows porting version 3.6 - no warnings
- Fixed all issues causing compiler warnings. Compilation
  is now clean under MS Visual Studio 2013 Community Edition.

+++ FASTCAP/savemat_mod.c
@@ -97,7 +97,7 @@ double *pimag;  /* pointer to imag data */
 	x.mrows = mrows;
 	x.ncols = ncols;
 	x.imagf = imagf;
-	x.namlen = strlen(pname) + 1;
+	x.namlen = (long) strlen(pname) + 1;
 	mn = x.mrows * x.ncols;
 
 	fwrite(&x, sizeof(Fmatrix), 1, fp);
@@ -144,7 +144,7 @@ int mn;				/* real #entries, this dump only */
 	  x.mrows = mrows;
 	  x.ncols = ncols;
 	  x.imagf = imagf;
-	  x.namlen = strlen(pname) + 1;
+	  x.namlen = (long) strlen(pname) + 1;
 	
 	  fwrite(&x, sizeof(Fmatrix), 1, fp);
 	  fwrite(pname, sizeof(char), (int)x.namlen, fp);
